Split up the forms
