+++ lib/pages/calendar.dart
@@ -1,4 +1,4 @@
-import 'package:executive_planner/backend/event_list.dart';
+import 'package:executive_planner/backend/events/event_list.dart';
 import 'package:executive_planner/widgets/bottom_nav_bar.dart';
 import 'package:executive_planner/widgets/event_list_display.dart';
 import 'package:flutter/material.dart';
